feat: add authorization and authentication stack and integrate into APIs
Add the Identity Pool for Federation with CDK, and add the Cognito user
pool which will be used for the reference application (later we will
add functionality/guidance in the IG for integrating with existing IdP).
Define default roles/groups for authorization.
Add positive and negative integration tests for testing identity pool/API Gateway
integration.