feat: add first-time federated users to the database
Add a function that runs pre-execution checks before all
lambda handlers, and call it in create-lambda handler.
Create 2 SSM parameters, one to store the user pool id,
and one to store the client id; these will be needed
to verify the idToken in the pre-checks.
Add all first time federated users to the DB when they
call a DEA API for the first time; call this code in
the lambda prechecks.