refactor: revoke old sessions for new login
1. If user a logins, calls an API, then logs in somewhere else and calls an API
then the second call will mark the first session as revoked and proceed. If the user
tries to use DEA with first set of credentials, then it will be blocked with a 412
REAUTHENTICATION ERROR
2. Refactor to use origin_jti instead of jti as tokenId for session, to ensure
that if a session is revoked, they cannot use the refresh token to get a new idtoken
they have to re-login (as stated by CJIS compliance). All tokens from the same authentication
session with cognito share an origin_jti
3. Remove IAM Authorizer from Refresh/Revoke Token
4. Fix broken E2E auth tests