chore: refactor stack for nested stack deployment in us-gov-east-1 du… (#264)
…e to Cognito inavailability

Gov Region us-gov-east-1 does NOT have Cognito therefore for that region we will launch the Cognito UserPool and IdPool in us-gov-west-1, and the rest of dea in us-gov-east-1. This will require DEA to be deployed in 3 stacks. MainStack: the main parent stack with various constructs Nested Stack: the auth stack with cognito and IAM, cognito is deployed in gov-west here for us-gov-east (IAM is global) Nested Stack: the stack that stores parameters for DEA. Split into a separate stack due to circular dependency.

This may lead to higher latencies, due to regional network travel.