feat: connect CognitoUserPool IdToken and IdentityId from IdentityPool together for users
To prevent users from using someone else's id token with their own credentials
(or the other way around), add IdPoolId to the user in the DDB table
and check that it matches for each API call. If it does not match, reject the call.

cr: https://code.amazon.com/reviews/CR-95295279