feat: add ability to integrate with Identity Center
Since Identity Center cannot send groups or custom attributes over the SAML assertion,
add a PreTokenGeneration Cognito Trigger when integrating with Identity Center. This trigger
lambda will query the identity store for the groups the member is a part of, and add it
to the id token, so authorization can happen as normal.

cr: https://code.amazon.com/reviews/CR-107849027